Merge the rails3 branch in preparation for Dashboard 2.0

+++ db/migrate/20100810011613_change_report_report_field_size_to_be_larger.rb
@@ -1,15 +1,22 @@
 class ChangeReportReportFieldSizeToBeLarger < ActiveRecord::Migration
-  def self.up
-    count = Report.delete_all(['length(reports.report) = ?', 65535])
-    if count > 0
-      say "WARNING: Deleted #{count} invalid reports that were truncated by the database due to an earlier bug. Please import your reports again (see README for details) if you wish to have these deleted reports re-added to your database correctly."
-    end
+  @postgres = ActiveRecord::Base.connection.adapter_name.downcase =~ /postgres/
 
-    say "Increasing column size of reports to avoid truncation problems:"
-    change_column( :reports, :report, :mediumtext, :limit => 16.megabytes )
+  def self.up
+    # data type 'text' is unlimited in postgres
+    unless @postgres
+      count = Report.delete_all(['length(reports.report) = ?', 65535])
+      if count > 0
+        say "WARNING: Deleted #{count} invalid reports that were truncated by the database due to an earlier bug. Please import your reports again (see README for details) if you wish to have these deleted reports re-added to your database correctly."
+      end
+
+      say "Increasing column size of reports to avoid truncation problems:"
+      change_column( :reports, :report, :mediumtext, :limit => 16.megabytes )
+    end
   end
 
   def self.down
-    change_column( :reports, :report, :text )
+    unless @postgres
+      change_column( :reports, :report, :text )
+    end
   end
 end
